setup workspace and related folders
