+++ apps/vet-server/README.md
@@ -1,0 +1,8 @@
+```
+npm install
+npm run dev
+```
+
+```
+open http://localhost:3000
+```
