+++ apps/vet-web/src/assets/PPEditorialNew-Bold.ttf
