+++ apps/vet-web/src/vite-env.d.ts
