+++ apps/vet-web/vite.config.ts
